Create options file
